Revert all CMOS changes back to (and including) 042d05f
This does a complete revert including all related future commits of 042d05f
which changes the CMOS date/time emulation to something that should more
closely match what real MS-DOS does. Though, this causes issues with some
software not working correctly anymore which all were bisected back to the
mentioned commit. The main reason for the revert is
https://github.com/joncampbell123/dosbox-x/issues/4725 (old NT-based OSes
not being able to properly load boot configuration changes like video driver
updates and such, making them hard to use properly, while DOSBox-X is able
to - while using a few quirks for the installation - perfectly run Windows
NT up to and including Windows XP in an emulated machine).

Since the upstream developer doesn't seem to really care about this
regression and having the need for an (emulated) NT machine, this revert is
the workaround for that issue.